v0.3.1
Changes:
-packet now moves from the center of the router to the center of another router
-better looking warnings
-length of the animation now depends on the length of the line connecting the 2 routers

+++ WebPacketSimulator.Windows/WpfRouter.cs
@@ -157,8 +157,8 @@ namespace WebPacketSimulator.Wpf
             for (int i = 0; i < routers.Count; i++)
             {
                 var router = routers[i];
-                int leftMargin = (int)(router.RouterImage.Margin.Left + moveAmmount.X);
-                int topMargin = (int)(router.RouterImage.Margin.Top + moveAmmount.Y);
+                double leftMargin = router.RouterImage.Margin.Left + moveAmmount.X;
+                double topMargin = router.RouterImage.Margin.Top + moveAmmount.Y;
                 router.RouterImage.Margin = new Thickness(leftMargin, topMargin, 0, 0);
             }
         }
